Edit loader for order
