+++ src/App.jsx
@@ -4,7 +4,8 @@ import Error from "./ui/Error";
 import Menu from "./features/menu/Menu";
 import { loader as menuLoader } from "./features/menu/MenuLoader";
 import Cart from "./features/cart/Cart";
-import CreateOrder, { action as createOrderAction } from "./features/order/CreateOrder";
+import CreateOrder from "./features/order/CreateOrder";
+import { action as createOrderAction } from "./features/order/Action";
 import Order from "./features/order/Order";
 import AppLayout from "./ui/AppLayout";
 import OrderLoader from "./features/order/OrderLoader";
@@ -32,7 +33,7 @@ const router = createBrowserRouter([
   },
   { path: '/order/:orderId',
     element: <Order />,
-    OrderLoader,
+    loader: OrderLoader,
     errorElement: <Error />,
   },
     ]
