+++ src/features/order/Action.jsx
@@ -1,0 +1,18 @@
+import { createOrder } from "../../services/apiRestaurant";
+import { redirect } from "react-router-dom";
+
+
+export async function action({ request }) {
+    const formData = await request.formData();
+    const data = Object.fromEntries(formData);
+  
+    const order = {
+      ...data,
+      cart: JSON.parse(data.cart),
+      priority: data.priority === "on",
+    };
+  
+    const newOrder = await createOrder(order);
+  
+    return redirect(`/order/${newOrder.id}`);
+  }
